// src/app/admin/consultation-medical/models/constance-physique.ts
export interface ConstancePhysique {

    id?: number;

    poids?: number;

    temperature?: number;

    tensionArterielS?: number;

    tensionArterielD?: number;

    frequenceR?: number;

    poul?: number;

    taille?: number;

    imc?: number;

    frequenceC?: number;

    autreConstances?: string;

}